# docs/swagger.yaml
definitions:
  MenuPayloads.ArticleBodyDetail:
    properties:
      article_body_paragraph:
        type: string
      article_image_content_path:
        type: string
    type: object
...
  MenuPayloads.ArticleSelectResponses:
    properties:
      article_body_content:
        items:
          $ref: '#/definitions/MenuPayloads.ArticleBodyDetail'
        type: array
      article_created_by_user_id:
        type: integer
      article_date_created:
        type: string
      article_header:
        type: string
      article_header_image_path:
        type: string
      article_id:
        type: integer
      is_bookmark:
        type: boolean
    type: object
...
  MenuPayloads.CalendarInsertPayload:
    properties:
      event_date:
        type: string
      event_name:
        type: string
      event_time_from:
        type: string
      event_time_to:
        type: string
      user_id:
        type: integer
    type: object
  MenuPayloads.CalendarUpdatePayload:
    properties:
      event_date:
        type: string
      event_id:
        type: integer
      event_name:
        type: string
      event_time_from:
        type: string
      event_time_to:
        type: string
      user_id:
        type: integer
    type: object
...
  entities.ArticleBodyEntities:
    properties:
      article_body_id:
        type: integer
      article_body_paragraph:
        type: string
      article_id:
        type: integer
      article_image_content_path:
        type: string
    type: object
  entities.ArticleEntities:
    properties:
      article_body:
        items:
          $ref: '#/definitions/entities.ArticleBodyEntities'
        type: array
      article_created_by_user_id:
        type: integer
      article_date_created:
        type: string
      article_header:
        type: string
      article_header_path_content:
        type: string
      article_id:
        type: integer
    type: object
  entities.Bookmark:
    properties:
      article:
        $ref: '#/definitions/entities.ArticleEntities'
      article_id:
        description: |-
          BookmarkTypeId int `gorm:"column:bookmark_type_id" json:"bookmark_type_id"`
          BookmarkType   BookmarkType
        type: integer
      bookmark_id:
        type: integer
      user:
        $ref: '#/definitions/entities.Users'
      user_id:
        type: integer
    type: object
  entities.UserDetail:
    properties:
      user_detail_id:
        type: integer
      user_gender:
        type: string
      user_height:
        type: number
      user_id:
        type: integer
      user_phone_number:
        description: UserProfileImage       string  `gorm:"column:user_profile_image"
          json:"user_profile_image"`
        type: string
      user_profile_description:
        type: string
      user_weight:
        type: number
    type: object
  entities.Users:
    properties:
      user_detail:
        allOf:
        - $ref: '#/definitions/entities.UserDetail'
        description: IsVIP        bool       `gorm:"column:is_vip" json:"is_vip"`
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
      user_password:
        type: string
    type: object
...
  payloads.LoginPaylods:
    properties:
      user_email:
        type: string
      user_password:
        type: string
    type: object
  payloads.RegisterPayloads:
    properties:
      user_email:
        type: string
      user_gender:
        type: string
      user_height:
        type: number
      user_name:
        type: string
      user_password:
        type: string
      user_phone_number:
        type: string
      user_weight:
        type: number
    type: object
  responses.ErrorResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  responses.StandarAPIResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: devin@gmail.com
    name: IoT Blynk Smart Agriculture Backend Service
    url: IoT
  description: IoT Blynk Smart Agriculture Backend Service
  license:
    name: IoT Blynk Smart Agriculture Backend Service
  termsOfService: http://swagger.io/terms/
  title: IoT Blynk Smart Agriculture Backend Service
  version: "1.0"
paths:
...
  /api/bookmark/{article_id}:
    delete:
      consumes:
      - application/json
      description: Remove Bookmark
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: integer
      - description: article_id
        in: query
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Remove Bookmark
      tags:
      - Bookmark
    post:
      consumes:
      - application/json
      description: Add New BookMark
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: integer
      - description: article_id
        in: query
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Add New BookMark
      tags:
      - Bookmark
  /api/bookmark/{article_type_id}:
    get:
      consumes:
      - application/json
      description: Get bookmark by Id
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: article_type_id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MenuPayloads.ArticleSelectResponses'
            type: array
      security:
      - BearerAuth: []
      summary: Get bookmark by Id
      tags:
      - Calendar
  /api/calendar:
    patch:
      consumes:
      - application/json
      description: Update Calendar
      parameters:
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.CalendarUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Update Calendar
      tags:
      - Calendar
    post:
      consumes:
      - application/json
      description: Create New Calendar
      parameters:
      - description: Insert Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.CalendarInsertPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Create New Calendar
      tags:
      - Calendar
  /api/calendar/by-user-id/:
    get:
      consumes:
      - application/json
      description: Get Calendar by Id
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
      security:
      - BearerAuth: []
      summary: Get Calendar by Id
      tags:
      - Calendar
  /api/calendar/delete/{calendar_id}:
    delete:
      consumes:
      - application/json
      description: Delete Calendar by Id
      parameters:
      - description: calendar_id
        in: path
        name: calendar_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
      security:
      - BearerAuth: []
      summary: Delete Calendar by Id
      tags:
      - Calendar
...
  /api/user/login2:
    post:
      consumes:
      - application/json
      description: Login by header
      parameters:
      - description: Insert Header Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payloads.LoginPaylods'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Login by header
      tags:
      - Auth
  /api/user/register:
    post:
      consumes:
      - application/json
      description: register by heade
      parameters:
      - description: Insert Header Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payloads.RegisterPayloads'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: register by heade
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"